+++ ClashFX/General/Utils/AppLogoTool.swift
@@ -117,6 +117,7 @@ enum AppLogoTool {
         }
         NSWorkspace.shared.noteFileSystemChanged(bundlePath)
         refreshIconAppearanceCache()
+        NotificationCenter.default.post(name: .appLogoDidChange, object: nil)
         return didSetIcon
     }