+++ ClashFX/Macro/Notification.swift
@@ -15,6 +15,7 @@ extension Notification.Name {
     static let systemNetworkStatusDidChange = Notification.Name("kSystemNetworkStatusDidChange")
     static let proxyMeneViewShowLeftPadding = Notification.Name("kProxyMeneViewShowLeftPadding")
     static let iCloudConfigStorageDidChange = Notification.Name("kICloudConfigStorageDidChange")
+    static let appLogoDidChange = Notification.Name("kAppLogoDidChange")
 
     static func proxyUpdate(for name: ClashProxyName) -> Notification.Name {
         return Notification.Name("kProxyUpdate_\(name)")